+++ .changeset/cool-lines-write.md
@@ -4,7 +4,7 @@
 
 Intersect
 
-Instead of `onIntersect` and `onExit` methods, the `Intersect` element now fires the `intersect` and `exit` custom events instead. If you were using these methods via JavaScript you'll need to update your code to listen for the custom event instead of the method. If you aren't using these methods via JavaScript, no changes are needed.
+Instead of `onIntersect` and `onExit` methods, the `Intersect` element now fires the `intersect` and `exit` custom events instead. If you were using these methods via JavaScript you'll need to update your code to listen for the custom event instead of using the methods. If you aren't using these methods via JavaScript, no changes are needed.
 
 ```diff
 - intersect.onIntersect(() => {
