+++ apps/docs/src/lib/code-controls.ts
@@ -20,7 +20,7 @@ export const codeControls: PluginSimple = (md: MarkdownIt) => {
 <div class="bg-base-800 rounded-none sm:rounded-md my-6 -mx-6 sm:mx-0 shadow-sm">
 	<div class="flex justify-between items-center pt-px px-4 sm:px-2 border-b border-base-700 gap-2">
 		<div class="font-mono px-2 text-base-200 text-sm">${lang}</div>
-		${Copy(escaped)}
+		${Share(escaped)}
 	</div>
 	${code}
 </div>
@@ -28,9 +28,9 @@ export const codeControls: PluginSimple = (md: MarkdownIt) => {
 	};
 };
 
-const Copy = (value: string) =>
+const Share = (value: string) =>
 	/* html */ `
-<drab-copy value="${value}">
+<drab-share value="${value}">
 	<button
 		data-trigger
 		type="button"
@@ -44,5 +44,5 @@ const Copy = (value: string) =>
 			<span class="icon-[lucide--clipboard-check]"></span>
 		</template>
 	</button>
-</drab-copy>
+</drab-share>
 `.trim();
