--- apps/docs/src/pages/docs/elements/copy/index.html
@@ -1,16 +1,0 @@
-<drab-copy value="value">
-	<button
-		data-trigger
-		type="button"
-		class="gap-2"
-		aria-label="copy value to clipboard"
-	>
-		<span>Copy</span>
-		<span data-content class="contents">
-			<span data-content class="icon-[lucide--clipboard-copy]"></span>
-		</span>
-		<template data-swap>
-			<span class="icon-[lucide--clipboard-check]"></span>
-		</template>
-	</button>
-</drab-copy>
