--- packages/drab/src/copy/define.ts
@@ -1,4 +1,0 @@
-import { define } from "../util/define.js";
-import { Copy } from "./index.js";
-
-define("drab-copy", Copy);
